Retrieve MAC address
This adds a function to retrieve the hardware address for one of the
machine's network interfaces, preferably eth0. If none is available, it
returns a fake address that is not a valid MAC address.

It also tries to determine if it is on an ODROID with no MAC address
tied to the hardware, and in that case it will also return a fake
address.

In any case, subsequent calls to the emtr_mac_gen() function are
guaranteed to be identical.

[endlessm/eos-sdk#438]