API to send queued data
This adds emtr_sender_send_queued_data() and its synchronous variation
which attempt to re-send all the data in the queue. Currently they are
not very smart about being failsafe: for example, the queue is emptied
and the items are sent one by one, being written back to the queue if
they fail.

Possible improvements: aborting the operation on the first failure, or
locking the queue to prevent race conditions.

[endlessm/eos-sdk#442]