Event Recorder Testing Constructor available
A testing constructor with modifiable parameters was needed for future
integration testing.  In order to have the customizable options, we
also needed the event recorder to have properties and a bunch more
private data instead of global constants.

The MachineIdProvider was created to give an abstraction for how the
UUIDs are stored and generated etc.  It is public API that may be
used by other objects but is currently only used by the Event
Recorder.

[endlessm/eos-sdk#1049]