DBus calls made
The DBus calls are being made synchronously from the event recorder. They are
calling a no-op DBus server for the time being, but we have a mock DBus
interface set up in python that is working for testing the validity of the data.
The event recorder still handles all of the logic that will eventually move to
the DBus server.

[endlessm/eos-sdk#1087]