Move functionality to client-side daemon
Rename EmtrMachineIdProvider to Emer... (EMER = Endless Metrics Event
Recorder) and move it to the daemon source code. Create a new directory
for testing daemon functionality. In the daemon tests, link the test
script directly with the corresponding C source file, since we are not
dealing with public API in a library.

Add an EmerDaemon object to the daemon. This object is the analogy of
the EmtrEventRecorder object in the library. However, it is not public
API so we can make a few simplifications: e.g. no need to provide a
..._get_default() singleton.

Refactor common functionality between the daemon and library into a
shared utility module.

(Devin & Philip)

[endlessm/eos-sdk#582]