Create a permissions provider
This adds an EmerPermissionsProvider to the daemon, not yet hooked up.
It reads a configuration file, $(sysconfdir)/eosmetrics-permissions.conf
by default. The file is a key-value file controlling whether the event
recorder server should record metrics or not (in the case that the user
has opted out.)

This code does not yet create the file.

(Devin & Philip)

[endlessm/eos-sdk#1324]