Monitor config file and reload or recreate on change
This monitors the /etc/eosmetrics-permissions.conf configuration file,
and reloads it on change or recreates it on delete. This functionality
was determined to be too flaky since the file monitor sends some
change events as a "delete" event immediately followed by a "create"
event; but here it is in case it is needed.

(Devin & Philip)

[endlessm/eos-sdk#1324]