Permissions provider controls whether to record
This hooks up the permissions provider to the daemon, by adding another
construct property to the daemon. The permission is read once at
construct time and subsequently whenever the permissions provider emits
a ::notify::daemon-enabled signal.

There is not a good way to test this yet since all the sending
functionality is in the daemon and can't yet be mocked out, but the
tests make sure that the property is read at least once after it is
set to FALSE.

[endlessm/eos-sdk#1325]