Make daemon-enabled property writable
The daemon will export an Enabled property on DBus through which the
first boot experience and other opt-out UIs can disable recording of
metrics.

This makes the EmerPermissionsProvider:daemon-enabled property writable
and the EmerDaemon:permissions-provider property readable (so that the
handler for DBus property changes can access its daemon-enabled
property directly.)

Setting the daemon-enabled property triggers an asynchronous write to
the config file.

[endlessm/eos-sdk#1326]