Can't requeue a modified message
Calling soup_session_requeue_message() on the message after we've
modified it fails, even though the message object is the same. Soup
treats it as a different message. Instead, we have to recreate a new
message and queue it with soup_session_queue_message(), specifying the
same callback and callback data.

[endlessm/eos-sdk#1361]