Abort test explicitly after timeout
Instead of calling g_main_loop_quit on timeout, explicitly call a
function that prints a message and asserts that it is not reached.

This is an attempt to solve the "Source ID 3 was not found when
attempting to remove it" failure on Jenkins. Perhaps the test is timing
out and the source is getting removed because it doesn't return TRUE.