Public getter for offset added.
There is now a way for outside classes to get the value of the
relative offset. This getter does not usually write to disk, unlike
the typical (and previously written) calls to update_offset().

A test was added to check the caching functionality of the PC w.r.t.
the relative offset. This test was not possible without the public
getter that was just introduced.
[endlessm/eos-sdk#1377]