Move event structs to util file.
These event structs are about to be used from the persistent cache, so it will
be valuable to share them.

Also add some memory management utilities for these structs and utilities to
convert them to GVariants. All of these utilities will be used from multiple
classes in the next commit.

[endlessm/eos-sdk#1419]