Add polkit authorization to enabling/disabling metrics
The server is set to require explicit authorization from Polkit
to access the new `SetEnabled` method that switches whether the
metrics server is enabled or not. The `enabled` property is now
set to read-only. Special cases for Polkit that authorize switching
metrics in the gnome-control-center and gnome-initial-setup are
contained in other commits in those repositories.
[endlessm/eos-sdk#1327]