Make the persistent cache API accept structs.
Currently we store lists of GVariants. This is problematic because GVariants are
immutable, and the persistent cache needs to mutate the GVariants before
persisting them.

Furthermore, the event recorder daemon does not use GVariants while metrics are
in the RAM buffers. It uses the structs that this change places in the
persistent cache API. Consequently, this change will simplify integration of the
persistent cache and the event recorder daemon.

[endlessm/eos-sdk#1419]