Add the metric request parser
Added the metric request parser. It accepts a raw hexadecimal
string that represents the serialized form of the metric
GVariant. The hexadecimal string must contain an even number
of characters and must not begin with "0x". The metric
request parser does not accept the raw binary representation
of the metric request, because the metric request is allowed
to have null bytes as part of the body. Null bytes are not
acceptable as command line input.

[endlessm/eos-sdk#1475]