Remove old-school metrics library code
This removes the unused code from the metrics library. Part of this
code ended up in the event recorder daemon in a different form. More of
it, such as the OS version provider, will be moved there in the future,
but for now I am deleting it because we are going to start API
versioning at the end of this release cycle (#418).

(Begun by Devin and finished by Philip)

[endlessm/eos-sdk#1024]