Integrated CacheVersionProvider with PC
The Daemon uses the PC which uses the CacheVersionProvider.
This cleaned up and simplified code in the PC. Testing code
now has to create its own (non-default) CacheVersionProvider
at least until we have a mock in its place.

Additionally, a bug was somehow uncovered where we were dependent
on NOT cleaning up a metrics testing file for our drain_when_empty
test to pass. The test was behaving appropriately but we needed
to add code to setup that created empty metrics files.
[endlessm/eos-sdk#1108]