Disable login tests when making Debian package.
We succeeded at disabling the login tests on Jenkins, but not on OBS. This
change disables the login tests on OBS as well. By login tests, we mean those
that use systemd-logind (and don't mock it out using Python's dbusmock).

This change is temporary until we port the systemd-logind tests to use Python's
dbusmock.

[endlessm/eos-shell#2870]