Permit metrics user and group to enable metrics.
Currently metrics remain disabled even when the user requests that they
be enabled in the FBE. This may, at least in part, be because the
metrics user and group are not listed in the policy kit file. This
change lists the metrics user and group in the policy kit file in hopes
that this will allow the metrics system to be enabled.

[endlessm/eos-sdk#1573]