Dont write to disk on store() always.
Previously, we were writing to our metafile more often than we needed
to. We were writing periodically to update our timestamps, and we were
writing every time store() was called.  This was not necessary and was
a performance/durability liability.

Instead, we only write to disk via metafile on store() when some form
of corruption is detected.  This required rewiring several unit tests
to use the public API getter of the boot offset instead of relying
on an implementation detail of store() that used to exist.
[endlessm/eos-sdk#1564]