Fixed testing segault from singleton.
Within the boot id provider, the singleton pattern was being used
unnecessarily which was harmless but for the unref() call in the
testing code. This would cause segfaults when the tests were run
on a machine actually using the daemon at the time.

This pattern causes complexity, and has been replaced by a normal
new() function for production and a new_full() for testing. There
is no get_default() anymore. I changed the name because by convention
the get_default() functions return singletons.
[endlessm/eos-sdk#1596]