Removed forward-declarations.
The table of contents is a maintenance burden because keeping it vertically
aligned is O(n) when we add a new function with a long name or a function
that returns or accepts something with a long type. This change removes the
table of contents, and shuffles functions around until they compile.

[endlessm/eos-sdk#1458]