Correct typos.
The output is a bit misleading in one of the error cases because the
error can be hit if the file isn't found or if permission is denied,
but the error only mentions the first case. Also, there is a $2
(second input) where there should be a &2 (stderr file), which causes
a crash when the script is passed an unexepected environment string.

[endlessm/eos-sdk#1576]