PC doesn't try to make files private.
The Persistent Cache was constantly vying for control of its files
against systemd-tempfiles. This resulted in excess writes. The
PC no longer uses G_FILE_CREATE_PRIVATE, which should allow for
permissions more permissive than 0600.

Additionally the metrics cache folder that is created (if it doesn't
exist when the PC is constructed) has its permissions changed to
allow for "all" users to have read permissions.
[endlessm/eos-sdk#1577]