Wait for mock DBus to come up
This introduces a small helper script that waits for the mock metrics
object to appear on the mock DBus. Previously we relied on "sleep 1s"
to be long enough for the service to come up, which could easily break
if the system load was high. This solution with the helper script also
avoids a sleep loop, preferring the GLib main loop instead.

(Philip & Devin)

[endlessm/eos-sdk#1168]