Wait for event recorder daemon to finish
In the test where the locally built event recorder daemon is launched,
we must wait for it to finish at the end of each test; and for good
measure, assert that it completed successfully.

Previously, it was possible that a new launch of the daemon would fail
because the old daemon had not terminated yet. In that case, the non-
zero exit code would not be caught either, so the test would still
pass.

[endlessm/eos-sdk#1565]