Ensure mock org.freedesktop.login1 service running
For tests that access the login manager, e.g. where the functionality
inhibits shutdown, we start a mock login manager service using
python-dbusmock. This is necessary because on Jenkins, in the chroot
environment, trying to access the logind service will bring up a new
instance of it, running in the chroot. That service won't quit, and so
the chroot's /sys will not get unmounted when the build is finished,
preventing all further builds unless manual action is taken.

For test-opt-out-integration.py, this is simple; we simply start the
mock logind service at the same time as the mock polkitd service.

We remove the instantiation of the mock service from the
launch-mock-dbus-tests.sh shell script, because for test-daemon, we
have to create the service from within the C code as in a later stage
we will have to interact with its stdout pipe.

For test-library, which previously used launch-mock-dbus-tests.sh, we
create a separate launcher shell script, with the previous contents of
launch-mock-dbus-tests.sh, which launches the mock event recorder
service.

[endlessm/eos-sdk#1565]