Re-enable asynchronous test.
test_permissions_provider_set_daemon_enabled_updates_config_file was
disabled because it fails on Jenkins with a timeout. Apparently we
don't receive the expected signals from GFileMonitor when the config
file is created and then modified. As a temporary (and flakey)
workaround we permit the test to proceed after the timeout rather than
failing it. In essence, we pretend we received any signals we are
waiting on after waiting on after 5 seconds. If the file still hasn't
been modified, the test will fail because it won't have the expected
contents. However, the test may fail when it should have passed in the
unlikely event that the file takes longer than 5 seconds to update.

[endlessm/eos-sdk#1429]