Remove event recorder API.
We are open-sourcing the event recorder API since it is used in mutter,
and we want to comply with mutter's GPL license. We are keeping the
event recorder server closed source, so the eos-metrics repository is
being renamed to eos-event-recorder-daemon. The event recorder API will
then be added to a new public repository named eos-metrics.

[endlessm/eos-sdk#2043]