Support starting before logind.
Previously we required the event recorder daemon to start after logind
so we could conveniently inhibit shutdown on startup. The
instrumentation daemon starts before logind to more conveniently record
logins. The instrumentation daemon should start after the event recorder
daemon so that it shuts down before the instrumentation daemon. This is
necessary so that the instrumentation daemon can record the shutdown
event in time for the event recorder daemon to persist it to disk.
However, this results in a systemd ordering cycle unless we break one of
the ordering dependencies. This change breaks this would-be cycle by
removing the requiremment that the event recorder daemon start after
logind. Instead of inhibiting shutdown on startup, the event recorder
daemon now inhibits shutdown on startup or once the login manager starts
up, whichever happens last.

[endlessm/eos-sdk#2906]