Warn when no valid boot offset is found.
Warn if there is no value with the group and key we expect for the boot
offset in the boot offset metadata file. Also warn if there a boot
offset in the metadata file, but its value is not an integer. In
general, we almost always want to log at least a warning when something
goes wrong rather than silently ignore it.

[endlessm/eos-sdk#1559]