Don't read from disk when boot offset is cached.
The metrics system should not do unnecessary I/O since it could slow
down user-facing operations. We are about to add more calls to the
function that updates the boot offset at shutdown, when performance is
an even greater concern. We want to have as much time as possible to
perform any necessary writes to disk before power is lost, and slow
shutdowns are user-antagonistic.

[endlessm/eos-sdk#1559]