Update persistent cache timestamps on finalize.
The absolute clock has a sneaky tendency to jump around while the
computer is on for a variety of reasons (e.g., NTP updates, users
manually setting the clock). Writing the current absolute and relative
time at the last possible moment gives very little time for any jumps of
this nature to throw off the metrics system's timekeeping algorithm.

[endlessm/eos-sdk#1559]