Permit NULL offset out parameter.
We are about to add a production call to
emer_persistent_cache_get_boot_time_offset that has no use for the
offset out parameter. Permit this parameter to be NULL, in which case
it simply isn't set. Update the persistent cache tests that don't need
the offset parameter to pass NULL rather than a dummy variable.

[endlessm/eos-sdk#1558]