Update timestamps on shutdown.
When the PrepareForShutdown signal is sent with parameter TRUE,
indicating that we are about to shutdown, update the relative and
absolute timestamps in the persistent cache's boot offset metadata
file. This makes the metrics' systems time-keeping algorithm more robust
in the event that the timestamp update that is triggered when the
persistent cache is finalized doesn't make it to disk before the system
shuts down.

[endlessm/eos-sdk#1558]