Remove conditional expression syntax.
Currently our Autoconf macro that checks whether the version of a Python
module is at least a certain value uses Python's conditional expression
syntax, which was introduced in Python 2.5. Being a version checking
macro, it should not assume which version of Python it has available to
it. Remove the conditional expression syntax since we didn't, strictly
speaking need it in the first place.

[endlessm/eos-sdk#3002]