Send metrics on first try.
Currently we don't send any metrics on the first try on a new image
because we notice that there is no network send file, write a default
send file and bail. Instead of bailing out, we now plow ahead and send
metrics regardless of whether we were able to find a network send file.

[endlessm/eos-sdk#3003]