Update timestamps after flushing to disk.
When the PrepareForShutdown signal is sent we have a limited amount of
time to perform any necessary operations and release the shutdown
inhibitor. We should perform the higher priority operations first on the
off chance that we don't have sufficient time to perform all necessary
work. In such a case, we would prefer to risk less accurate timing
information than lose events, so we update the timestamps in the boot
offset metadata file only after we have flushed any events to disk.

[endlessm/eos-sdk#3018]