Update timestamps before releasing inhibitor.
Currently, in emer_daemon_finalize, we release the shutdown inhibitor
after flushing the persistent cache but before updating the timestamps
in the persistent cache's boot offset metadata file. The point of
inhibiting shutdown is to give us enough time to perform any necessary
disk writes before the system shuts down, so we now perform all
necessary disk writes before releasing the shutdown inhibitor. This is
is unlikely to be a problem in practice because we only expect
emer_daemon_finalize to be called after the PrepareForShutdown signal
would be sent.

[endlessm/eos-sdk#3018]