Move update_boot_offset to initialization.
We now call update_boot_offset once an hour to make sure that we keep
the relative and absolute timestamps up-to-date in the event of an
unexpected power loss. There is no longer a need to call
update_boot_offset every time we store events. We can simply do this
once at initialization to handle the case where events are stored before
the first hour elapses.

[endlessm/eos-sdk#3243]