Support specifying max cache size via file.
Read the maximum persistent cache size out of the key file
/etc/metrics/cache-size.conf rather than hard-coding it. Introduce a
cache size provider to abstract away the details of reading from the key
file. This functionality will allow us to increase the maximum size of
the persistent cache for the clients we send on the imminent bus trip
without affecting any other clients.

[endlessm/eos-sdk#3151]