Add uploading_enabled field to permissions file.
Support enabling and disabling of network uploads of metrics data via a
new variable in /etc/metrics/eos-metrics-permissions.conf. This
functionality is important for the computers going on the bus trip.
Those computers will be configured with large persistent caches, and we
don't want the metrics system wracking up a large data bill for an
unsuspecting user who connects the computer to the network via
tethering or a 3G dongle.

[endlessm/eos-sdk#3151]