Correct sed syntax.
sed currently crashes because it expects the value that comes after the
equals sign to be a suffix to use for a backup of the file being
modified in place. We were passing the script itself where the suffix
belonged. We now pass the script as a separate argument as sed expects.

[endlessm/eos-sdk#3124]