Simplify updating of network requests.
Our network payloads include the time the request was sent, but
GVariants are immutable, so when we retry a network request we create a
completely new GVariant from the old one. Rather than extracting arrays
as GVariantIters and then converting them to GVariantBuilders, we should
use @ to extract them as GVariants and use these directly.

[endlessm/eos-sdk#3168]