Emit upload-finished signal when failing fast.
Consider the case where the timer starts an upload, then the
eos-upload-metrics script is run twice in a row, and finally the metrics
system is disabled. The timer-driven upload will finish, but without
this change only the task corresponding to the first run of the
eos-upload-metrics script will be dequeued, and the other task can end
up leaked. This change ensures that tasks can't end up stuck in the
queue indefinitely.

[endlessm/eos-sdk#3243]