Standardize persistent cache format.
In the event that we ever need to retrieve data from a persistent cache
via USB or some other manual mechanism, it will be a lot less hassle to
interpret that data if the persistent cache format is not
machine-dependent. Eliminate the storage format's dependencies on the
endianness and word size of the machine. Increment the cache version
number so that any data stored in the persistent cache is wiped when
this upgrade is received.

[endlessm/eos-sdk#3254]