Properly handle references on auxiliary payloads.
This change fixes a longstanding memory leak, a longstanding crash, and
a longstanding bug where the persistent cache would claim to have stored
events even when the write to disk failed. In order to properly handle
floating references on auxiliary payloads in
emer_persistent_cache_store_metrics, we should not take ownership of
auxiliary payloads of events that end up not being stored in cache. In
particular, events may not be stored as requested if either they don't
fit or the write fails. Since we can't know whether a disk write will
fail until we try it, we are forced to deep copy the payloads, only
ref_sinking and unreffing the original payloads if the corresponding
events fit in the cache and are successfully written to disk.

[endlessm/eos-sdk#3254]