Remove dangerous error handling.
Writing to disk in response to errors that can only result if the system
isn't behaving as intended is risky. We shouldn't pretend we understand
how to get ourselves out of states that we never should have entered in
the first place. Even if we are correct such error handling code is
difficult to test.

[endlessm/eos-sdk#599]