Fix crash at shutdown.
The metrics instrumentation daemon expects to be able to send messages
to the event recorder daemon as long as the event recorder daemon is
alive. The event recorder daemon is a D-Bus service, so it can only
receive messages while D-Bus is around to deliver them. Therefore,
instruct systemd to start D-Bus before and shut it down after the event
recorder daemon.

[endlessm/eos-sdk#3848]