Clear all the metrics data when corrupt data is found in the cache
The binary format stored in the persistent cache is very sensitive to
unexpected changes, meaning that once we get into a state where the
server reads corrupt data out of it (e.g. zero-sized elements), it will
stay stuck into that situation unless it somehow manages to find the
next valid chunk of data in the circular file, which is not trivial.

For now, to prevent this situation where a device remains stuck
forever, we detect the issue accordingly while reading from the
circular file and reset all its contents when that happens, so
that at least the metrics can be used again afterwards.

https://phabricator.endlessm.com/T12739