In case of invalid data, try to keep as much valid information as possible
The current implementation of the persistent cache is a (FIFO) circular
file where we basically read data from a HEAD up to SIZE (up to a maximum
size, as configured). In this situation, if we read valid data for a while
after HEAD but then find invalid elements before reading all the SIZE bytes,
we could still use the good ones we read and simply discard the rest.

So, this patch makes sure that we do only read valid elements from the
circular file and that the invalid bits get ignored, so that we don't
raise an G_IO_ERROR_INVALID_DATA error from emer_persistent_cache_read()
unless it's strictly needed (i.e. when not having detected it sooner).

https://phabricator.endlessm.com/T12739