Send a metric event when invalid or corrupt data is found in the cache
In order to detect when this kind of situation happens, and how often,
report two new metrics events right after dealing with the data stored
in the cache, so that we can analyze them accordingly.

At the moment, two new metrics events are defined:

* CACHE_IS_CORRUPT_EVENT_ID: "d84b9a19-9353-73eb-70bf-f91a584abcbd"
* CACHE_HAS_INVALID_ELEMENTS_EVENT_ID: "cbfbcbdb-6af2-f1db-9e11-6cc25846e296"

The first one won't include any payload and will simply serve to inform
when the whole cache has been purged, in the event that some malformed
data had made it to the file.

The second one will include a '(tt)' payload with the number of valid
records read from the cache and the number of bytes read in total.

https://phabricator.endlessm.com/T12739