permissions-provider: write config from idle, not thread
GKeyFile is not thread-safe, so it is not appropriate to call
g_key_file_save_to_file() from another thread.

Instead, schedule an idle callback to rewrite the file. This keeps the
SetEnabled() D-Bus method non-blocking, and will allow multiple updates
to be coalesced into a single write. In principle we could use
g_key_file_to_data() + g_file_replace_contents_async() to avoid blocking
the main thread on writing to disk but I don't think this is necessary
for now: this is hardly a frequently-used API.