main: set uploading-enabled when SetEnabled is called
With this change, the (read-only) Enabled D-Bus property reflects only
whether metrics are collected; but updating it also changes whether
those metrics are submitted over the network. This allows those settings
to have separate defaults, but to be controlled in unison by the FBE and
other UI.

Only three combinations are actually meaningful: don't collect; collect
but don't upload; collect and upload. With this change, it's only
possible to set the first and third combination via the D-Bus API,
whereas previously it was impossible to change whether uploading is
enabled via the D-Bus API.

https://phabricator.endlessm.com/T14065