Change default configuration to "collect, don't upload"
With the previous defaults, metrics collected during first boot would be
dropped, even if the user consented to metrics in the FBE. This is
problematic because the boot-time metrics include crucial facts like the
OS version are not submitted the first time an Endless OS installation
is used. This also prevented us collecting metrics from live sessions,
since the "is this a live session?" metric is also reported at boot
time, and every boot is the first boot.

There are only three meaningful combinations of these two flags:

1. enabled=false, uploading_enabled is irrelevant (old default)
2. enabled=true,  uploading_enabled=false (new default)
3. enabled=true,  uploading_enabled=true

With this new default, boot-time metrics will be collected but not
submitted. Once the FBE is completed:

* if the user opted out of metrics collection, SetEnabled(False) will
  move us to configuration 1 and the metrics gathered so far will be
  discarded
* if the user opted into metrics collection, SetEnabled(True) will move
  us to configuration 3 and the boot-time metrics will be submitted as
  soon as a network connection is available.

https://phabricator.endlessm.com/T14065