avoid race in tests/test-opt-out-integration.py
This change relies on the fact that, after reading it once, the daemon
does not monitor or reload the permissions config file - it only writes
to it after a change has been made. So, for the purposes of this test,
in order to check that the configuration file appears on disk in
various states, we can wait for it to be created, check the contents,
then delete it. The next time it re-appears on disk will be after
another state change caused by the test.