daemon: discard pending events on SetEnabled(False)
At present, if you disable metrics collection, any cached but
unsubmitted events are preserved indefinitely. If you later re-enable
metrics, these old events will be exhumed and submitted along with
newly-collected ones.

In particular, by the time the FBE offers the user the chance to opt out
of metrics, boot-time events have already been collected (but not
submitted). If you opt out during the FBE, you might be surprised to
find cached events in /var/cache/metrics; if you later opt back in, you
might also be surprised to see them submitted.

This patch discards all pending events (in-memory or on-disk) when
metrics are disabled. If a request is already in flight, it's too late
to cancel, but it will not be retried.

https://phabricator.endlessm.com/T16713