daemon: ref NetworkCallbackData.daemon
In finish_network_callback, the EmerDaemon object was previously not
guaranteed to live long enough to emit a signal on it.
NetworkCallbackData.upload_task owns a ref to it, but the task was
unreffed before the signal was emitted.

In practice, something else does always hold a reference to the
EmerDaemon, but it's clearer to make this explicit.