cache-size-provider: replace class with a function
If you tune out all the GObject noise, EmerCacheSizeProvider is just a
function that reads an integer from a file (creating the file if it's
missing), exactly once. So we can delete all the class code in favour of
a function that returns a guint64, and pass that guint64 to
EmerPersistentCache.

This simplification makes clearer an existing bug: if the file can't be
parsed, the cache size returned is 0. When that's fed to
EmerCircularFile via EmerPersistentCache, the on-disk cache will be
truncated to 0 bytes, and no events will ever be persisted to disk
again.

While we're here: I think it's incorrect to round-trip the test data
through GKeyFile before reading it. A more honest, and simpler, test is
to write the test data directly to disk; this will also allow us to test
the case where the data cannot be parsed by GKeyFile.