Fix distcheck when run as non-root
When running `make distcheck`, $(libdir) expands to `/usr/local/lib`.
This is not typically writable by non-root users, so the build fails:

     /usr/bin/mkdir -p '/usr/local/lib/systemd'
    /usr/bin/mkdir: cannot create directory ‘/usr/local/lib’: Permission denied
    Makefile:2196: recipe for target 'install-systemdsystemunitDATA' failed
    make[3]: *** [install-systemdsystemunitDATA] Error 1

Instead we should force this directory to be somewhere within the
prefix. Unlike the last few patches, this actually is not a regression
in https://phabricator.endlessm.com/T19953!