Use AX_COMPILER_FLAGS to enable many, many warnings
The warnings it enables are a superset of Emmanuele's list of flags.
Whether or not they are errors is controlled by a configure parameter,
with sensible defaults based on whether we're building from a git
checkout. During distcheck, all warnings are explicitly fatal.

Previously, these warnings (some of which are fatal) were only enabled
when running `make distcheck`, which at least I evidently do not
routinely do during development.