daemon: Fix potential unaligned accesses on ARM
`-Wcast-align` has helpfully been pointing these out in the ARM compiler
output for a while.

The circular buffer consists of a load of length-encoded buffers, which
don’t appear to have any guaranteed alignment. The length is stored as a
guint64, which was being read directly out of the circular buffer — if
unaligned, this would cause SIGBUS on ARM.

Fix that by reading it out of the circular buffer using memcpy().

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T26444