collector: Fixes to get EOS version
We need to check we are running on EOS and not other distro before
comparing versions. Also, move the method out of the class to a global
function, since it is not really specific to the OfflineMetrics class.

Finally, let's check the version before instantiating the uploader
object so exit fast if we won't be able to do any work anyway.

https://phabricator.endlessm.com/T30684