collector: Improve arg parsing and usage description
This adds a list of possible choices and a default value to the 'action'
argument. This way argparse performs the argument validation for us,
exiting the program if the argument is not one of the specified choices.

It also improves the usage message and also move the description string
to a separate variable, to make it easier to read the source.

Finally, it gets the program location from sys.path, so we don't need to
process sys.argv[0].

https://phabricator.endlessm.com/T30684