Get script location from sys.path and move it after argument parsing
sys.path has a list of directories where python should look for modules,
the first one being the directory where the script is running from.

Moving 'location' after argument parsing brings it closer to where it is
used, making it easier to find when reading through the code. Also,
having argument parsing being the first thing in main makes it work
better as documentation for someone reading the source.