collector: Wrap all calls to systemctl into one function
Instead of calling subprocess.run(systemctl, ...) everywhere, let's
create a function for that (and drop os.system which is obsolete).

Also, we don't need to check if the unit is running before trying to
stop it, but we should only try to start it after the operation if it
was running before.

https://phabricator.endlessm.com/T30684