collector: Small logic fix and refactor attributes and variables
Do not use 'is True' or 'is False': these are comparing if a object is
the same as the object True or the object False. To compare the True of
False values in python, one should use 'if expression' or 'if not
expression'.

Use the same label for location, storage_path and
offline_metrics_usbdisk_node: these three variables are representing the
same data in different layers. We can use the same label everywhere to
make it easier to read. Also rename offline_machine_data_dir to
target_dir, to improve readability by shortening some lines.

This commit also moves the offline_metrics_usbdisk_node to the parent
class, since it is used by both OfflineMetricsCollector and
OfflineMetricsUploader.

Finally, it drop the unused attribute disk_free_space_in_bytes, avoids
defining the 'eos-metrics-data' name more than once and removes some
extra blank lines.

https://phabricator.endlessm.com/T30684