collector: Check for tracking_id in the constructor
This reduces code duplication. Also we don't need to check if machine-id
exists -- it is a strong requirement of systemd-based systems, so we can
safely assume it does.

This commit also closes the {tracking,machine}_id file immediately after
reading it, since there is no need to keep it open any longer.

Finally, it also improves the comment in copy_tracking_id to explain why
we can't use shutil.copy.

https://phabricator.endlessm.com/T30684