collector: Do not silently remove existing metrics
Let's not remove existing metrics from the USB device without warning
the user, as this may lead to data loss. The upload operation removes
the data after uploading, so if it is around most likely it hasn't been
uploaded yet (perhaps because the upload operation failed without the
user noticing it?).

If we want to make it easier for the user, we could request for
confirmation here (but this commit does not implement that).

https://phabricator.endlessm.com/T30684