collect: Do not silently remove existing metrics
Let's not remove existing metrics without warning the user, as this may
lead to data loss. If we want to make it easier for the user, we may
request for confirmation. But I'm fine with the current solution, since
the upload operation should remove the data on upload. If the data is
present, most likely the upload operation failed without the user
noticing it, or the user have not tried to upload.