collector: Check feasibility before stopping unit
Check if there is any metrics data to upload and try to create target
dir before trying to stop the metrics service unit. This way we only
interfere with the daemon's life if we are sure we are going to do some
actual work.

Also rename create_folder_for_machine to create_machine_dir, which is a
bit more concise, and improve error message and exit with a success
status when there is no data to upload.

https://phabricator.endlessm.com/T30684