collector: Refactor the code that spawns the metrics daemon
This breaks long lines and name variables in a way that is easier to
read, and only calls mkdtemp for the top-level temp directory, outside
of the per-machine loop.

Also, it drops calling killall in a separate process in favor of using
the Popen object's terminate() method, and drops calling rm in a
separate process in favor of shutil.rmtree.

Finally, do not suppress the daemon's output, as this will be useful for
debugging.

https://phabricator.endlessm.com/T30684